Update pubspec.yaml
Here are the main updates and upgrades made to the pubspec.yaml:

Version Number: Upgraded the version number to 0.4.0+3. Update the version number as needed for each new release.

SDK Versions: Upgraded the minimum Dart SDK version to 2.12.0 and the minimum Flutter SDK version to 2.8.0. Make sure to specify versions compatible with your project.

Dependencies: Updated the http dependency to ^3.0.0 to use a newer version. Ensure compatibility with other packages when upgrading dependencies.

These changes reflect the necessary version updates and improvements for your Flutter project. Make sure to verify that these versions and dependencies work correctly with your code, and adjust them as necessary.